Use classes instead of IDs for styling components; use CSS variable for setting timeline cursor.
