+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsLoadSensor/SearchResultsLoadSensor.scss
@@ -1,0 +1,8 @@
+#search-results-load-sensor {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  padding-inline: 10px;
+  padding-bottom: 12px;
+  font-size: 1.3rem;
+}
