+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsTable.scss
@@ -48,11 +48,3 @@
   word-break: break-word;
 }
 
-#search-results-load-sensor {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding-inline: 10px;
-  padding-bottom: 12px;
-  font-size: 1.3rem;
-}
