+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsTimeline.scss
@@ -1,9 +1,5 @@
-#timeline-chart {
-  cursor: crosshair;
+.timeline-chart {
+  cursor: var(--timeline-chart-cursor, crosshair);
   height: 100px;
   max-height: 100px;
-}
-
-.timeline-chart-cursor-disabled {
-  cursor: wait !important;
 }
